use a single VideoModelMock definition
